Add convert functions

+++ AdditionCalculator/AdditionCalculator/Program.cs
@@ -7,11 +7,11 @@
             Console.WriteLine("Hello, World!");
 
             Console.WriteLine("Enter the First Number");
-            int number1 = Int32.Parse(Console.ReadLine());
+            double number1 = Double.Parse(Console.ReadLine());
             Console.WriteLine("Enter the Second Number");
-            int number2 = Int32.Parse(Console.ReadLine());
+            double number2 = Double.Parse(Console.ReadLine());
 
-            int number3 = number1 + number2;
+            double number3 = number1 + number2;
             Console.WriteLine("The sum is: " + number3);
 
             //Alternate way of writing it
